refactor(server): extract import session boundary
