+++ packages/server/src/server/agent/create-agent-title.ts
@@ -1,0 +1,37 @@
+import { MAX_EXPLICIT_AGENT_TITLE_CHARS } from "./agent-title-limits.js";
+
+const MAX_INITIAL_AGENT_TITLE_CHARS = Math.min(60, MAX_EXPLICIT_AGENT_TITLE_CHARS);
+
+function deriveInitialAgentTitle(prompt: string): string | null {
+  const firstContentLine = prompt
+    .split(/\r?\n/)
+    .map((line) => line.trim())
+    .find((line) => line.length > 0);
+  if (!firstContentLine) {
+    return null;
+  }
+  const normalized = firstContentLine.replace(/\s+/g, " ").trim();
+  if (!normalized) {
+    return null;
+  }
+  const clamped = normalized.slice(0, MAX_INITIAL_AGENT_TITLE_CHARS).trim();
+  return clamped.length > 0 ? clamped : null;
+}
+
+export function resolveCreateAgentTitles(options: {
+  configTitle?: string | null;
+  initialPrompt?: string | null;
+}): { explicitTitle: string | null; provisionalTitle: string | null } {
+  const explicitTitle =
+    typeof options.configTitle === "string" && options.configTitle.trim().length > 0
+      ? options.configTitle.trim()
+      : null;
+  const trimmedPrompt = options.initialPrompt?.trim();
+  const provisionalTitle =
+    explicitTitle ?? (trimmedPrompt ? deriveInitialAgentTitle(trimmedPrompt) : null);
+
+  return {
+    explicitTitle,
+    provisionalTitle,
+  };
+}
