+++ packages/server/src/server/agent/import-sessions.ts
@@ -1,18 +1,25 @@
 import type { z } from "zod";
+import type { Logger } from "pino";
 import type { ProviderDefinition } from "./provider-registry.js";
-import type { AgentManager } from "./agent-manager.js";
+import type { AgentManager, ManagedAgent } from "./agent-manager.js";
 import type { AgentStorage, StoredAgentRecord } from "./agent-storage.js";
 import type {
   AgentPersistenceHandle,
   AgentProvider,
+  AgentSessionConfig,
+  AgentTimelineItem,
   PersistedAgentDescriptor,
 } from "./agent-sdk-types.js";
+import { scheduleAgentMetadataGeneration } from "./agent-metadata-generator.js";
+import { resolveCreateAgentTitles } from "./create-agent-title.js";
+import { unarchiveAgentState } from "./agent-prompt.js";
 import { toRecentProviderSessionDescriptorPayload } from "./agent-projections.js";
 import type {
   FetchRecentProviderSessionsRequestMessage,
   ImportAgentRequestMessageSchema,
   RecentProviderSessionDescriptorPayload,
 } from "../../shared/messages.js";
+import type { WorkspaceGitService } from "../workspace-git-service.js";
 
 type ImportAgentRequestMessage = z.infer<typeof ImportAgentRequestMessageSchema>;
 
@@ -47,6 +54,23 @@ export interface ListImportableProviderSessionsInput {
 export interface ListImportableProviderSessionsResult {
   entries: RecentProviderSessionDescriptorPayload[];
   filteredAlreadyImportedCount: number;
+}
+
+export interface ImportProviderSessionInput {
+  request: NormalizedImportAgentRequest;
+  agentManager: AgentManager;
+  agentStorage: AgentStorage;
+  workspaceGitService?: Pick<WorkspaceGitService, "resolveRepoRoot">;
+  paseoHome?: string;
+  logger: Logger;
+  deps?: {
+    scheduleAgentMetadataGeneration?: typeof scheduleAgentMetadataGeneration;
+  };
+}
+
+export interface ImportProviderSessionResult {
+  snapshot: ManagedAgent;
+  timelineSize: number;
 }
 
 // COMPAT(import-agent-request-v1): accept legacy {provider, sessionId} shape
@@ -118,6 +142,99 @@ export async function listImportableProviderSessions(
   return { entries, filteredAlreadyImportedCount };
 }
 
+export async function importProviderSession(
+  input: ImportProviderSessionInput,
+): Promise<ImportProviderSessionResult> {
+  const { provider, providerHandleId, cwd, labels } = input.request;
+  const descriptor = await input.agentManager.findPersistedAgent(provider, providerHandleId);
+  if (!descriptor && provider === "opencode" && !cwd) {
+    throw new Error(
+      "OpenCode sessions require --cwd when the session cannot be found in persisted agents",
+    );
+  }
+
+  const handle = descriptor
+    ? applyImportCwdOverride(descriptor.persistence, cwd)
+    : buildImportPersistenceHandle({ provider, providerHandleId, cwd });
+  const overrides = cwd ? ({ cwd } satisfies Partial<AgentSessionConfig>) : undefined;
+
+  await unarchiveAgentByHandle(input.agentStorage, input.agentManager, handle);
+  const snapshot = await input.agentManager.resumeAgentFromPersistence(
+    handle,
+    overrides,
+    undefined,
+    {
+      labels,
+    },
+  );
+  await unarchiveAgentState(input.agentStorage, input.agentManager, snapshot.id);
+  await input.agentManager.hydrateTimelineFromProvider(snapshot.id);
+  await applyImportedAgentTitle({
+    snapshot,
+    agentManager: input.agentManager,
+    workspaceGitService: input.workspaceGitService,
+    paseoHome: input.paseoHome,
+    logger: input.logger,
+    scheduleAgentMetadataGeneration:
+      input.deps?.scheduleAgentMetadataGeneration ?? scheduleAgentMetadataGeneration,
+  });
+
+  return {
+    snapshot,
+    timelineSize: input.agentManager.getTimeline(snapshot.id).length,
+  };
+}
+
+async function unarchiveAgentByHandle(
+  agentStorage: AgentStorage,
+  agentManager: AgentManager,
+  handle: AgentPersistenceHandle,
+): Promise<void> {
+  const records = await agentStorage.list();
+  const matched = records.find(
+    (record) =>
+      record.persistence?.provider === handle.provider &&
+      record.persistence?.sessionId === handle.sessionId,
+  );
+  if (!matched) {
+    return;
+  }
+  await unarchiveAgentState(agentStorage, agentManager, matched.id);
+}
+
+async function applyImportedAgentTitle(input: {
+  snapshot: ManagedAgent;
+  agentManager: AgentManager;
+  workspaceGitService?: Pick<WorkspaceGitService, "resolveRepoRoot">;
+  paseoHome?: string;
+  logger: Logger;
+  scheduleAgentMetadataGeneration: typeof scheduleAgentMetadataGeneration;
+}): Promise<void> {
+  const initialPrompt = getFirstUserMessageText(input.agentManager.getTimeline(input.snapshot.id));
+  if (!initialPrompt) {
+    return;
+  }
+
+  const { explicitTitle, provisionalTitle } = resolveCreateAgentTitles({
+    configTitle: input.snapshot.config.title,
+    initialPrompt,
+  });
+  if (!explicitTitle && provisionalTitle) {
+    await input.agentManager.setTitle(input.snapshot.id, provisionalTitle);
+  }
+
+  input.scheduleAgentMetadataGeneration({
+    agentManager: input.agentManager,
+    agentId: input.snapshot.id,
+    cwd: input.snapshot.cwd,
+    workspaceGitService: input.workspaceGitService,
+    initialPrompt,
+    explicitTitle,
+    paseoHome: input.paseoHome,
+    logger: input.logger,
+  });
+}
+
 function parseRecentProviderSessionsSince(since: string | undefined): number | null {
   if (!since) {
     return null;
@@ -127,6 +244,54 @@ function parseRecentProviderSessionsSince(since: string | undefined): number | n
     throw new ImportSessionsRequestError("invalid_since", "Invalid recent provider sessions since");
   }
   return timestamp;
+}
+
+function buildImportPersistenceHandle(input: {
+  provider: AgentProvider;
+  providerHandleId: string;
+  cwd?: string;
+}): AgentPersistenceHandle {
+  const cwd = input.cwd ?? process.cwd();
+  return {
+    provider: input.provider,
+    sessionId: input.providerHandleId,
+    nativeHandle: input.providerHandleId,
+    metadata: {
+      provider: input.provider,
+      cwd,
+    },
+  };
+}
+
+function applyImportCwdOverride(
+  handle: AgentPersistenceHandle,
+  cwd: string | undefined,
+): AgentPersistenceHandle {
+  if (!cwd) {
+    return handle;
+  }
+
+  return {
+    ...handle,
+    metadata: {
+      ...handle.metadata,
+      provider: handle.provider,
+      cwd,
+    },
+  };
+}
+
+function getFirstUserMessageText(timeline: readonly AgentTimelineItem[]): string | null {
+  for (const item of timeline) {
+    if (item.type !== "user_message") {
+      continue;
+    }
+    const text = item.text.trim();
+    if (text) {
+      return text;
+    }
+  }
+  return null;
 }
 
 async function collectImportedProviderSessionHandles(
