+++ packages/server/src/server/session.test.ts
@@ -14,12 +14,9 @@ import type {
   AgentClient,
   AgentMode,
   AgentModelDefinition,
-  AgentTimelineItem,
   ListModesOptions,
   ListModelsOptions,
-  PersistedAgentDescriptor,
 } from "./agent/agent-sdk-types.js";
-import type { ManagedAgent } from "./agent/agent-manager.js";
 import type { ProviderDefinition } from "./agent/provider-registry.js";
 import { ProviderSnapshotManager } from "./agent/provider-snapshot-manager.js";
 import type { SessionOptions } from "./session.js";
@@ -70,8 +67,6 @@ interface SessionHandlerInternals {
   handleCheckoutPullRequest(params: unknown): Promise<unknown>;
   handleCheckoutPushRequest(params: unknown): Promise<unknown>;
   handleCheckoutStatusRequest(params: unknown): Promise<unknown>;
-  handleFetchRecentProviderSessions(params: unknown): Promise<unknown>;
-  handleImportAgentRequest(params: unknown): Promise<unknown>;
   describeWorkspaceRecord(...args: unknown[]): Promise<WorkspaceDescriptorPayload>;
   describeWorkspaceRecordWithGitData(...args: unknown[]): Promise<WorkspaceDescriptorPayload>;
   handleValidateBranchRequest(params: unknown): Promise<unknown>;
@@ -932,262 +927,6 @@ function createWorkspaceGitSnapshot(
 
 afterEach(() => {
   vi.clearAllMocks();
-});
-
-describe("session agent import", () => {
-  test("filters already imported provider sessions by native handle", async () => {
-    const messages: unknown[] = [];
-    const cwd = "/tmp/import-filter";
-    const importedDescriptor: PersistedAgentDescriptor = {
-      provider: "claude",
-      sessionId: "provider-session-imported",
-      cwd,
-      title: "Already imported",
-      lastActivityAt: new Date("2026-04-30T12:00:00.000Z"),
-      persistence: {
-        provider: "claude",
-        sessionId: "provider-session-imported",
-        nativeHandle: "provider-native-imported",
-      },
-      timeline: [{ type: "user_message", text: "already imported" }],
-    };
-    const visibleDescriptor: PersistedAgentDescriptor = {
-      provider: "claude",
-      sessionId: "provider-session-visible",
-      cwd,
-      title: "Still importable",
-      lastActivityAt: new Date("2026-04-30T11:00:00.000Z"),
-      persistence: {
-        provider: "claude",
-        sessionId: "provider-session-visible",
-        nativeHandle: "provider-native-visible",
-      },
-      timeline: [{ type: "user_message", text: "still importable" }],
-    };
-    const agentManager = {
-      listAgents: vi.fn(() => [
-        {
-          provider: "claude",
-          persistence: {
-            provider: "claude",
-            sessionId: "stored-session-id",
-            nativeHandle: "provider-native-imported",
-          },
-        } as ManagedAgent,
-      ]),
-      listImportablePersistedAgents: vi
-        .fn()
-        .mockResolvedValue([importedDescriptor, visibleDescriptor]),
-      subscribe: vi.fn(() => () => {}),
-    };
-    const agentStorage = {
-      list: vi.fn().mockResolvedValue([]),
-    };
-    const session = createSessionForTest({ messages });
-    Object.assign(session, { agentManager, agentStorage });
-
-    await asSessionInternals(session).handleFetchRecentProviderSessions({
-      type: "fetch_recent_provider_sessions_request",
-      requestId: "recent-provider-sessions",
-      cwd,
-      providers: ["claude"],
-      limit: 15,
-    });
-
-    expect(agentManager.listImportablePersistedAgents).toHaveBeenCalledWith({
-      limit: 200,
-      providerFilter: new Set(["claude"]),
-      cwd,
-    });
-    expect(messages).toContainEqual(
-      expect.objectContaining({
-        type: "fetch_recent_provider_sessions_response",
-        payload: expect.objectContaining({
-          requestId: "recent-provider-sessions",
-          filteredAlreadyImportedCount: 1,
-          entries: [
-            expect.objectContaining({
-              providerId: "claude",
-              providerHandleId: "provider-native-visible",
-              title: "Still importable",
-            }),
-          ],
-        }),
-      }),
-    );
-  });
-
-  test("imports by provider handle id and hydrates title metadata from the first user message", async () => {
-    const messages: unknown[] = [];
-    const cwd = "/tmp/imported-agent";
-    const timeline: AgentTimelineItem[] = [
-      { type: "user_message", text: "Trace recent provider sessions\n\nkeep it tight" },
-      { type: "assistant_message", text: "I will inspect the provider listing." },
-    ];
-    const snapshot = {
-      id: "00000000-0000-4000-8000-000000000633",
-      provider: "custom-codex",
-      cwd,
-      capabilities: TEST_CAPABILITIES,
-      config: { provider: "custom-codex", cwd },
-      createdAt: new Date("2026-04-30T00:00:00.000Z"),
-      updatedAt: new Date("2026-04-30T00:00:00.000Z"),
-      availableModes: [],
-      currentModeId: null,
-      pendingPermissions: new Map(),
-      bufferedPermissionResolutions: new Map(),
-      inFlightPermissionResponses: new Set(),
-      pendingReplacement: false,
-      persistence: {
-        provider: "custom-codex",
-        sessionId: "thread-imported",
-        nativeHandle: "provider-thread-imported",
-        metadata: { provider: "custom-codex", cwd },
-      },
-      historyPrimed: true,
-      lastUserMessageAt: null,
-      attention: { requiresAttention: false },
-      foregroundTurnWaiters: new Set(),
-      finalizedForegroundTurnIds: new Set(),
-      unsubscribeSession: null,
-      internal: false,
-      labels: {},
-      lifecycle: "closed",
-      session: null,
-      activeForegroundTurnId: null,
-    } satisfies ManagedAgent;
-    const agentManager = {
-      listAgents: vi.fn(() => []),
-      subscribe: vi.fn(() => () => {}),
-      findPersistedAgent: vi.fn().mockResolvedValue({
-        provider: "custom-codex",
-        sessionId: "thread-imported",
-        cwd,
-        title: null,
-        firstPromptPreview: "Trace recent provider sessions",
-        lastPromptPreview: "Trace recent provider sessions",
-        lastActivityAt: new Date("2026-04-30T00:00:00.000Z"),
-        persistence: snapshot.persistence,
-      }),
-      resumeAgentFromPersistence: vi.fn().mockResolvedValue(snapshot),
-      hydrateTimelineFromProvider: vi.fn().mockResolvedValue(undefined),
-      getTimeline: vi.fn().mockReturnValue(timeline),
-      setTitle: vi.fn().mockResolvedValue(undefined),
-      notifyAgentState: vi.fn(),
-    };
-    const agentStorage = {
-      list: vi.fn().mockResolvedValue([]),
-      get: vi.fn().mockResolvedValue(null),
-    };
-    const session = createSessionForTest({ messages });
-    Object.assign(session, { agentManager, agentStorage });
-
-    await asSessionInternals(session).handleImportAgentRequest({
-      type: "import_agent_request",
-      providerId: "custom-codex",
-      providerHandleId: "provider-thread-imported",
-      cwd,
-      requestId: "import-thread",
-    });
-
-    expect(agentManager.findPersistedAgent).toHaveBeenCalledWith(
-      "custom-codex",
-      "provider-thread-imported",
-    );
-    expect(agentManager.hydrateTimelineFromProvider).toHaveBeenCalledWith(snapshot.id);
-    expect(agentManager.setTitle).toHaveBeenCalledWith(
-      snapshot.id,
-      "Trace recent provider sessions",
-    );
-    expect(agentMetadataMocks.scheduleAgentMetadataGeneration).toHaveBeenCalledWith(
-      expect.objectContaining({
-        agentManager,
-        agentId: snapshot.id,
-        cwd,
-        initialPrompt: "Trace recent provider sessions\n\nkeep it tight",
-        explicitTitle: null,
-      }),
-    );
-  });
-
-  test("sets a provisional title and schedules auto-title generation from the first hydrated user message", async () => {
-    const messages: unknown[] = [];
-    const cwd = "/tmp/imported-agent";
-    const timeline: AgentTimelineItem[] = [
-      { type: "user_message", text: "Investigate flaky checkout status\n\ninclude logs" },
-      { type: "assistant_message", text: "I will inspect the checkout flow." },
-    ];
-    const snapshot = {
-      id: "00000000-0000-4000-8000-000000000632",
-      provider: "codex",
-      cwd,
-      capabilities: TEST_CAPABILITIES,
-      config: { provider: "codex", cwd },
-      createdAt: new Date("2026-04-30T00:00:00.000Z"),
-      updatedAt: new Date("2026-04-30T00:00:00.000Z"),
-      availableModes: [],
-      currentModeId: null,
-      pendingPermissions: new Map(),
-      bufferedPermissionResolutions: new Map(),
-      inFlightPermissionResponses: new Set(),
-      pendingReplacement: false,
-      persistence: {
-        provider: "codex",
-        sessionId: "thread-imported",
-        nativeHandle: "thread-imported",
-        metadata: { provider: "codex", cwd },
-      },
-      historyPrimed: true,
-      lastUserMessageAt: null,
-      attention: { requiresAttention: false },
-      foregroundTurnWaiters: new Set(),
-      finalizedForegroundTurnIds: new Set(),
-      unsubscribeSession: null,
-      internal: false,
-      labels: {},
-      lifecycle: "closed",
-      session: null,
-      activeForegroundTurnId: null,
-    } satisfies ManagedAgent;
-    const agentManager = {
-      listAgents: vi.fn(() => []),
-      subscribe: vi.fn(() => () => {}),
-      findPersistedAgent: vi.fn().mockResolvedValue(null),
-      resumeAgentFromPersistence: vi.fn().mockResolvedValue(snapshot),
-      hydrateTimelineFromProvider: vi.fn().mockResolvedValue(undefined),
-      getTimeline: vi.fn().mockReturnValue(timeline),
-      setTitle: vi.fn().mockResolvedValue(undefined),
-      notifyAgentState: vi.fn(),
-    };
-    const agentStorage = {
-      list: vi.fn().mockResolvedValue([]),
-      get: vi.fn().mockResolvedValue(null),
-    };
-    const session = createSessionForTest({ messages });
-    Object.assign(session, { agentManager, agentStorage });
-
-    await asSessionInternals(session).handleImportAgentRequest({
-      type: "import_agent_request",
-      provider: "codex",
-      sessionId: "thread-imported",
-      cwd,
-      requestId: "import-thread",
-    });
-
-    expect(agentManager.setTitle).toHaveBeenCalledWith(
-      snapshot.id,
-      "Investigate flaky checkout status",
-    );
-    expect(agentMetadataMocks.scheduleAgentMetadataGeneration).toHaveBeenCalledWith(
-      expect.objectContaining({
-        agentManager,
-        agentId: snapshot.id,
-        cwd,
-        initialPrompt: "Investigate flaky checkout status\n\ninclude logs",
-        explicitTitle: null,
-      }),
-    );
-  });
 });
 
 describe("session provider refresh cwd routing", () => {
